Добавлены managed биндинги для blob-ресурсов рендера
+ Расширены IRenderingFacade/INativeRenderingApi методами Create*FromBlob и DestroyResource
+ Реализованы P/Invoke и NativeRuntime для renderer_create_*_from_blob/renderer_destroy_resource
+ Переведены Mesh/Texture/Material handles на ulong для соответствия C ABI (uint64)
+ Обновлены Noop/Native stubs и добавлены тесты для interop и ресурсного API

+++ engine/managed/src/Engine.Core/Handles/MaterialHandle.cs
@@ -4,13 +4,13 @@ public readonly record struct MaterialHandle
 {
     public static MaterialHandle Invalid => default;
 
-    public MaterialHandle(uint value)
+    public MaterialHandle(ulong value)
     {
         HandleGuards.RequireNonZero(nameof(value), value);
         Value = value;
     }
 
-    public uint Value { get; }
+    public ulong Value { get; }
 
     public bool IsValid => Value != 0;
