+++ engine/managed/tests/Engine.Tests/Rendering/UiDrawCommandTests.cs
@@ -13,7 +13,7 @@ public sealed class UiDrawCommandTests
     {
         var command = new UiDrawCommand(new TextureHandle(55), 1, 2, 3, 4);
 
-        Assert.Equal((uint)55, command.Texture.Value);
+        Assert.Equal((ulong)55, command.Texture.Value);
         Assert.Equal((uint)1, command.VertexOffset);
         Assert.Equal((uint)2, command.VertexCount);
         Assert.Equal((uint)3, command.IndexOffset);
